requirements.txt
streamlit==1.5.0
pandas==1.3.3
datetime==4.3
matplotlib==3.4.2
numpy==1.25.0  
Pillow==6.2.2